requirements.txt
django
biopython
reportlab
plotly
pandas
